requirements.txt
Django
Markdown
Pillow
psycopg2
pytz
djangorestframework